WIP totally jacked collision.
Welllllllll that doesn't work for shit. It (seemingly) randomly lets the
ball bounce when the paddle is on the other side of the screen. I got
nothin'. Here's the original commit message I wrote when I thought this
worked, in a simpler time like 10 minutes ago.

-----------

Partially working smarter collision detection.

It turns out that this is the hardest problem in computer science. Who
knew. This is busted, as you can see in the TODO. It randomly (but I'm
sure not actually random) said there was a y-axis collision. Gameplay
dictates that you can play while ignoring all collisions on that axis
(except for the walls, which can be detected simply and separately).
You'll only notice this is jacked if the ball passes through the
top/bottom of a paddle. It'll look dumb, but won't affect the score
since that is a point lost or scored regardless.